[starts #163025615] combine signup and login tests in one file for easier testing -- write tests for login endpoint
